Update generate.php with compression
Generatedata sometimes provides files that are several MB in size. In order to save on bandwidth costs, data files over 500 KB will be compressed to zip. 100000 rows of data consisting of about 5 columns of data is about 6 MB uncompressed, and a little over 2 MB when not. That's over 60% bandwidth saved.

It would be great if we could add the compress option to the UI so that users would have a choice.